Some stylistic changes.

+++ src/com/codlex/thermocycler/logic/bath/sensors/TemperatureSensor.java
@@ -16,6 +16,6 @@ public class TemperatureSensor {
 	}
 
 	public float getTemperature() {
-		return this.sensor.getValue().floatValue();
+		return this.sensor.getValue();
 	}
 }
